requirements.txt
flask
flask_cors
pdfplumber
python-docx
python-pptx
nltk
requests
gunicorn